refactor: 调整函数位置并用上常量
